Creating core now
Creates a core. Deletes existing core, if any. I'd like to do the delete
a different way. I have to see if Vagrant allows provisioner steps that
are only run when called explicitly, not when doing a normal provision.

Changes: Start Solr before creating core. Separate Solr command to create
the core. Core is like an instance of Solr, when you're not in cloud mode.
Collection is the name when it's in cloud mode.