Broken -- The crawl fails
The crawl fails because some files in the linkdb don't get created. There
were all sorts of learning. Creating the cloud version (two nodes), you
have to make sure that both nodes are configured correctly. I was getting
404 errors sometimes about the "/update" action, but that was maybe because
the second node wasn't configured for Nutch.

Nutch is also a bit finicky about the URL filter and the URL itself. Put
the "/" at the end of the name. The regex-urlfilter.txt has to have the
right number of backslashes.

I still don't know why the solr server works and the vagrant machine
doesn't. I tried all the variations I could think of to find the
difference.