Minimalist working except for Nutch dedup bug.
This is back to a more minimal deploy of Solr. I also refactored the
whole thing so, for example, the .profile is set up early, when you
need it for debugging.

I can't figure out why the dedup error happens in the Vagrant machine,
but not my full-up VM on the server. The VM on the server only has 1 GB
of memory, just like the Vagrant machine, but the server has swap.