Add a custom crawl bash file.
I still don't know why the VM works and the Vagrant machine doesn't,
but I tracked down on Vagrant that it's the way it does the link
inversion inside the iterations that seems to cause the problems.

I replaced the crawl bash file that comes in the Nutch distribution
with one that does the inversion and deduplication at the end, outside
the loop.