Solr startup hopefully working now.
I had to set a special condition on the upstart solr.conf file so that
it wouldn't run until after the /vagrant drive was mounted.

Then I had to start manually at initial provisioning time, since the script
isn't in place yet when the directory is mounted.

Finally, the guard to ensure that the script is only run when /vagrant is
mounted stopped the script from running when used manually. I took it out,
but there's probably a more general solution for upstart in this case.