Abandon this attempt.
This commit is the start of a boxing exercise. But I realized that having
the config files in the /vagrant directory is sort of incompatible with
the idea of boxing, at least to the level that I understand it.

I could put all the config files on the VM, but you can't ln into the VM
(at least without doing some other stuff I don't know how to do).

Saving this branch for posterity.